Blocks: Merge Provider with BlockEdit
