+++ blocks/block-edit/context.js
@@ -1,52 +1,13 @@
 /**
  * External dependencies
  */
-import { Component, createContext, createHigherOrderComponent } from '@wordpress/element';
+import { createContext, createHigherOrderComponent } from '@wordpress/element';
 
 const { Consumer, Provider } = createContext( {
 	isSelected: true,
 } );
 
-/**
- * A Higher Order Component used to be provide a context for BlockEdit
- * component.
- *
- * @param {Component} OriginalComponent Component to wrap.
- *
- * @return {Component} Component with a BlockEdit context set.
- */
-export const withBlockEditContextProvider = createHigherOrderComponent( ( OriginalComponent ) => {
-	return class ComponentWithProvider extends Component {
-		constructor( props ) {
-			super( props );
-			this.state = {
-				context: {
-					isSelected: props.isSelected,
-				},
-			};
-		}
-
-		static getDerivedStateFromProps( nextProps, prevState ) {
-			if ( nextProps.isSelected === prevState.context.isSelected ) {
-				return null;
-			}
-
-			return {
-				context: {
-					isSelected: nextProps.isSelected,
-				},
-			};
-		}
-
-		render() {
-			return (
-				<Provider value={ this.state.context }>
-					<OriginalComponent { ...this.props } />
-				</Provider>
-			);
-		}
-	};
-}, 'withBlockEditContextProvider' );
+export { Provider as BlockEditContextProvider };
 
 /**
  * A Higher Order Component used to inject BlockEdit context to the
