+++ blocks/block-edit/index.js
@@ -19,9 +19,18 @@ import {
 	getBlockDefaultClassName,
 	hasBlockSupport,
 } from '../api';
-import { withBlockEditContextProvider } from './context';
+import { BlockEditContextProvider } from './context';
 
 export class BlockEdit extends Component {
+	constructor( props ) {
+		super( props );
+		this.state = {
+			context: {
+				isSelected: props.isSelected,
+			},
+		};
+	}
+
 	getChildContext() {
 		const {
 			id: uid,
@@ -35,6 +44,18 @@ export class BlockEdit extends Component {
 				'capabilities',
 				'unfiltered_html',
 			], false ),
+		};
+	}
+
+	static getDerivedStateFromProps( nextProps, prevState ) {
+		if ( nextProps.isSelected === prevState.context.isSelected ) {
+			return null;
+		}
+
+		return {
+			context: {
+				isSelected: nextProps.isSelected,
+			},
 		};
 	}
 
@@ -60,12 +81,14 @@ export class BlockEdit extends Component {
 		// For backwards compatibility concerns adds a focus and setFocus prop
 		// These should be removed after some time (maybe when merging to Core)
 		return (
-			<Edit
-				{ ...this.props }
-				className={ className }
-				focus={ isSelected ? {} : false }
-				setFocus={ noop }
-			/>
+			<BlockEditContextProvider value={ this.state.context }>
+				<Edit
+					{ ...this.props }
+					className={ className }
+					focus={ isSelected ? {} : false }
+					setFocus={ noop }
+				/>
+			</BlockEditContextProvider>
 		);
 	}
 }
@@ -76,7 +99,6 @@ BlockEdit.childContextTypes = {
 };
 
 export default compose( [
-	withBlockEditContextProvider,
 	withFilters( 'blocks.BlockEdit' ),
 	withSelect( ( select ) => ( {
 		postType: select( 'core/editor' ).getEditedPostAttribute( 'type' ),
